--- a/Makefile
+++ b/Makefile
@@ -2,7 +2,7 @@ PREFIX = /usr/local
 MANPREFIX = ${PREFIX}/share/man
 FUSE_UFS_FLAGS = -p fuse-ufs --ignore-rust-version --no-default-features -F $$(uname)
 
-BENCH = cargo bench -p rufs
+BENCH = cargo bench -p rufs --bench bench
 
 SRC != find rufs/src fuse-ufs/src -name '*.rs'
 
@@ -33,6 +33,9 @@ bench:
 	${BENCH} --no-default-features -F dcache > bench.dcache
 	${BENCH} --no-default-features -F bcache > bench.bcache
 	${BENCH} > bench.cache
+
+flame:
+	${BENCH} -F 'bcache' -- --profile-time 5
 
 fuz:
 	mkdir -p fuzz/corpus/ufs/