TODO:
- [x] Implement benchmarking
- [ ] Inode LruCache (see [lru](https://docs.rs/lru))
- [ ] Find other performance improvements

Original performance numbers (on a Ryzen 9 7900):
| Test | Variant              | Time    | Throughput  |
|------|----------------------|---------|-------------|
| open |                      | 20.2 us |             |
| read | 1MiB                 | 27.8 ms | 983 MiB/s   |
| read | 64KiB                | 28.7 ms | 955 MiB/s   |
| read | 16KiB                | 30.6 ms | 896 MiB/s   |
| read | 4KiB                 | 118 ms  | 232 MiB/s   |
| read | 512B                 | 934 ms  | 29 MiB/s    |
| find | direct               | 610 us  | 3.1 Mfile/s |
| find | lookup               | 1.77 ms | 1.1 Mfile/s |
| find | lookup+stat          | 109 ms  | 17 Kfile/s  |
| find | lookup+stat+read-64K | 193 ms  | 9.8 Kfile/s |
| find | lookup+stat+read-16K | 202 ms  | 9.3 Kfile/s |
| find | lookup+stat+read-4K  | 455 ms  | 4.2 Kfile/s |
| find | lookup+stat+read-512 | 2.9 s   | 652 file/s  |