@realchonk 2025-02-16
Performance numbers with an inode cache:
| Test | Variant              | Time    | Throughput  |
|------|----------------------|---------|-------------|
| open |                      | 20.2 us |             |
| read | 1MiB                 | 26.7 ms | 1023 MiB/s  |
| read | 64KiB                | 27.4 ms | 998 MiB/s   |
| read | 16KiB                | 27.5 ms | 994 MiB/s   |
| read | 4KiB                 | 106 ms  | 257 MiB/s   |
| read | 512B                 | 826 ms  | 33 MiB/s    |
| find | direct               | 441 us  | 4.3 Mfile/s |
| find | lookup               | 924 us  | 2.1 Mfile/s |
| find | lookup+stat          | 75 ms   | 25 Kfile/s  |
| find | lookup+stat+read-64K | 156 ms  | 12 Kfile/s  |
| find | lookup+stat+read-16K | 158 ms  | 12 Kfile/s |
| find | lookup+stat+read-4K  | 400 ms  | 4.7 Kfile/s |
| find | lookup+stat+read-512 | 2.6 s   | 718 file/s  |