@realchonk 2025-02-16
Performance numbers with:
- Inode cache
- redundant `read_inode()` removal

| Test | Variant              | Time    | Throughput  |
|------|----------------------|---------|-------------|
| open |                      | 19.7 us |             |
| read | 1MiB                 | 26 ms   | 1029 MiB/s  |
| read | 64KiB                | 26.6 ms | 1004 MiB/s  |
| read | 16KiB                | 26.3 ms | 1015 MiB/s  |
| read | 4KiB                 | 104 ms  | 264 MiB/s   |
| read | 512B                 | 797 ms  | 34 MiB/s    |
| find | direct               | 429 us  | 4.4 Mfile/s |
| find | lookup               | 885 us  | 2.1 Mfile/s |
| find | lookup+stat          | 73 ms   | 26 Kfile/s  |
| find | lookup+stat+read-64K | 155 ms  | 12 Kfile/s  |
| find | lookup+stat+read-16K | 157 ms  | 12 Kfile/s  |
| find | lookup+stat+read-4K  | 380 ms  | 5.0 Kfile/s |
| find | lookup+stat+read-512 | 2.4 s   | 783 file/s  |