@realchonk 2025-02-16
Performance numbers with:
- inode cache
- redundant `read_inode()` removal
- block cache

# With block cache
| Test | Variant              | Time    | Throughput  |
|------|----------------------|---------|-------------|
| open |                      | 23 us   |             |
| read | 1MiB                 | 26.1 ms | 1023 MiB/s  |
| read | 64KiB                | 26.4 ms | 1010 MiB/s  |
| read | 16KiB                | 27.0 ms | 1012 MiB/s  |
| read | 4KiB                 | 99 ms   | 276 MiB/s   |
| read | 512B                 | 747 ms  | 37 MiB/s    |
| find | direct               | 457 us  | 4.1 Mfile/s |
| find | lookup               | 913 us  | 2.1 Mfile/s |
| find | lookup+stat          | 70 ms   | 27 Kfile/s  |
| find | lookup+stat+read-64K | 153 ms  | 12 Kfile/s  |
| find | lookup+stat+read-16K | 155 ms  | 12 Kfile/s  |
| find | lookup+stat+read-4K  | 367 ms  | 5.1 Kfile/s |
| find | lookup+stat+read-512 | 2.3 s   | 808 file/s  |