@realchonk 2025-02-16
Performance numbers with:

- inode cache
- redundant read_inode() removal
-  block cache
- directory entry cache

| Test | Variant              | Time    | Throughput  |
|------|----------------------|---------|-------------|
| open |                      | 23 us   |             |
| read | 1MiB                 | 26.8 ms | 1019 MiB/s  |
| read | 64KiB                | 27.5 ms | 996 MiB/s   |
| read | 16KiB                | 27.9 ms | 983 MiB/s   |
| read | 4KiB                 | 102 ms  | 270 MiB/s   |
| read | 512B                 | 776 ms  | 35 MiB/s    |
| find | direct               | 461 us  | 4.1 Mfile/s |
| find | lookup               | 661 us  | 2.9 Mfile/s |
| find | lookup+stat          | 68 ms   | 28 Kfile/s  |
| find | lookup+stat+read-64K | 152 ms  | 12 Kfile/s  |
| find | lookup+stat+read-16K | 155 ms  | 12 Kfile/s  |
| find | lookup+stat+read-4K  | 375 ms  | 5.1 Kfile/s |
| find | lookup+stat+read-512 | 2.4 s   | 795 file/s  |