@realchonk 2025-02-16
Summary so far
| Test | Variant              | Baseline               | icache                 | dcache                 | bcache                 | icache+dcache          |
|------|----------------------|------------------------|------------------------|------------------------|------------------------|------------------------|
| open |                      | 20.7 us                | 20.3 us                | 20.6 us                | 22.4 us                | 20.9 us                |
| read | 1MiB                 | 26.0 ms / 1029 MiB/s   | 24.9 ms / 1074 MiB/s   | 24.3 ms / 1099 MiB/s   | 27.2 ms / 1005 MiB/s   | 25.6 ms / 1044 MiB/s   |
| read | 64KiB                | 26.8 ms / 1020 MiB/s   | 25.1 ms / 1065 MiB/s   | 25.0 ms / 1071 MiB/s   | 28.1 ms / 975 MiB/s    | 26.2 ms / 1021 MiB/s   |
| read | 16KiB                | 28.8 ms / 952 MiB/s    | 25.3 ms / 1055 MiB/s   | 27.6 ms / 990 MiB/s    | 29.9 ms / 915 MiB/s    | 26.3 ms / 1016 MiB/s   |
| read | 4KiB                 | 109 ms / 251 MiB/s     | 99.1 ms / 276 MiB/s    | 105 ms / 262 MiB/s     | 109 ms / 251 MiB/s     | 103 ms / 265 MiB/s     |
| read | 512B                 | 863 ms / 31.7 MiB/s    | 763 ms / 35.8 MiB/s    | 828 ms / 33.0 MiB/s    | 847 ms / 32.3 MiB/s    | 777 ms / 35.2 MiB/s    |
| find | direct               | 430 us / 4.41 Mfile/s  | 431 us / 4.40 Mfile/s  | 421 us / 4.50 Mfile/s  | 443 us / 4.28 Mfile/s  | 439 us / 4.32 Mfile/s  |
| find | lookup               | 1.22 ms / 1.56 Mfile/s | 876 us / 2.16 Mfile/s  | 760 us / 2.50 Mfile/s  | 1.20 ms / 1.58 Mfile/s | 639 us / 2.97 Mfile/s  |
| find | lookup+stat          | 77.8 ms / 24.4 Kfile/s | 77.0 ms / 24.6 Kfile/s | 72.6 ms / 26.1 Kfile/s | 75 ms / 25.2 Kfile/s   | 70.1 ms / 27.0 Kfile/s |
| find | lookup+stat+read-64K | 162 ms / 11.7 Kfile/s  | 148 ms / 12.8 Kfile/s  | 147 ms / 12.9 Kfile/s  | 160 ms / 11.9 Kfile/s  | 147 ms / 12.9 Kfile/s  |
| find | lookup+stat+read-16K | 170 ms / 11.2 Kfile/s  | 150 ms / 12.6 Kfile/s  | 155 ms / 12.2 Kfile/s  | 166 ms / 11.4 Kfile/s  | 150 ms / 12.6 Kfile/s  |
| find | lookup+stat+read-4K  | 411 ms / 4.61 Kfile/s  | 370 ms / 5.13 Kfile/s  | 377 ms / 5.03 Kfile/s  | 405 ms / 4.68 Kfile/s  | 373 ms / 5.09 Kfile/s  |
| find | lookup+stat+read-512 | 2.62 s / 722 file/s    | 2.36 s / 804 file/s    | 2.50 s / 759 file/s    | 2.59 s / 732 file/s    | 2.48 s / 763 file/s    |

Conclusions: 
- bcache is useless (or badly optimized)
- icache+dcache is king