@realchonk 2025-02-16
@asomers What do you think about this? Do you have any suggestions on what could improve the performance?